refactor:[ADD]LectureCRU body:Lecture관련 기능 상수 추가 및 카테고리별 조회 booter: #10
