+++ HH99SpringLv3/src/main/java/com/sparta/hh99springlv3/domain/lecture/entity/Category.java
@@ -5,8 +5,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class LCareer {
-    private String spring;
-    private String react;
-    private String node;
+public enum Category {
+    SPRING("SPRING"),REACT("REACT"), NODE("NODE"),;
+    private final String category;
 }
