+++ HH99SpringLv3/src/main/java/com/sparta/hh99springlv3/domain/lecture/repository/LectureRepository.java
@@ -1,10 +1,11 @@
 package com.sparta.hh99springlv3.domain.lecture.repository;
 
+import com.sparta.hh99springlv3.domain.lecture.entity.Category;
 import com.sparta.hh99springlv3.domain.lecture.entity.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface LectureRepository extends JpaRepository<Lecture,Long> {
-    List<Lecture> findByCategory(String category);
+    List<Lecture> findByCategory(Category category);
 }
